refactor(suggestion): refactor suggestion extension into smaller chunks and fix abort controller bug
